pyproject.toml
[project]
name = "open_ithageneia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-environ~=0.12.0",
    "django~=6.0.0",
    "django-vite~=3.1.0",
    "inertia-django~=1.2.0",
    "django-allauth~=65.14.0",
    "whitenoise~=6.11.0",
    "gunicorn~=23.0.0",
    "pillow~=12.1.0",
    "django-import-export~=4.4.0",
    "openpyxl~=3.1.5",
]
